KEYCLOAK-4750 resolving the promise with an undefined value to proceed to access denied with an invalid realm key
When a user has an invalid realm key, the token correctly fails verification (verify.verify)
but then in the validateGrant method the rejection throws the error and then the chain
continues. The user is granted access. By resolving the promise with an undefined token,
the user will be denied access with the 403 expected.

Tests can be written with the expected 403 in the keycloak-nodejs-connect code base.